# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sentinel file intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    get:
      operationId: search
      tags: [search]
      summary: Unified search across findings — combines a natural-language query with explicit, editable filters
      description: >
        `q` (if present) is run through the local NL interpreter first; any explicit filter
        params override/extend the interpreted ones. Every call is recorded in search history.
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: path
          in: query
          schema:
            type: string
        - name: extension
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: aiCategory
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: riskLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: minSizeBytes
          in: query
          schema:
            type: integer
        - name: maxSizeBytes
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: scanId
          in: query
          schema:
            type: integer
        - name: duplicatesOnly
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: recordHistory
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResults"
  /search/history:
    get:
      operationId: listSearchHistory
      tags: [search]
      summary: Recent search history, most recent first
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search history
          content:
            application/json:
              schema:
                type: object
                required: [history]
                properties:
                  history:
                    type: array
                    items:
                      $ref: "#/components/schemas/SearchHistoryEntry"
    delete:
      operationId: clearSearchHistory
      tags: [search]
      summary: Clear all search history
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                type: object
                required: [cleared]
                properties:
                  cleared:
                    type: integer
  /search/saved:
    get:
      operationId: listSavedSearches
      tags: [search]
      summary: List saved searches
      responses:
        "200":
          description: Saved searches
          content:
            application/json:
              schema:
                type: object
                required: [savedSearches]
                properties:
                  savedSearches:
                    type: array
                    items:
                      $ref: "#/components/schemas/SavedSearch"
    post:
      operationId: createSavedSearch
      tags: [search]
      summary: Save the current query + filters under a name
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSearchInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
  /search/saved/{id}:
    patch:
      operationId: updateSavedSearch
      tags: [search]
      summary: Rename or update the filters of a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveSearchInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedSearch"
        "404":
          description: Not found
    delete:
      operationId: deleteSavedSearch
      tags: [search]
      summary: Delete a saved search
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted:
                    type: boolean
...
components:
  schemas:
...
    Finding:
      type: object
      required: [id, scanId, type, path, name, extension, sizeBytes, findingStatus, riskLevel, reviewStatus, reason, createdAt]
      properties:
        id:
          type: integer
        scanId:
          type: integer
        type:
          type: string
          enum: [empty_folder, zero_byte, idlk_file, locked_file, installer, archive, large_file, duplicate]
        path:
          type: string
        name:
          type: string
        extension:
          type: string
        sizeBytes:
          type: integer
        hash:
          type: ["string", "null"]
        duplicateGroupHash:
          type: ["string", "null"]
        findingStatus:
          type: string
          enum: [safe_delete, review, duplicate, ignored]
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
          description: "Heuristic risk level, display-only — never drives automatic action"
        reviewStatus:
          type: string
          enum: [new, reviewed, accepted, rejected, ignored, quarantined]
          description: "Findings review workflow state — independent of findingStatus"
        reviewedAt:
          type: ["string", "null"]
        reason:
          type: string
        fileCreatedAt:
          type: ["string", "null"]
          description: "Filesystem creation timestamp captured at scan time"
        fileModifiedAt:
          type: ["string", "null"]
          description: "Filesystem modification timestamp captured at scan time"
        createdAt:
          type: string
        aiCategory:
          type: ["string", "null"]
          description: "High-level AI category (e.g. Legal, Tax, Photography, Software)"
        aiSubcategory:
          type: ["string", "null"]
          description: "Optional finer-grained classification within the category"
        aiConfidence:
          type: ["integer", "null"]
          minimum: 0
          maximum: 100
          description: "AI classification confidence 0–100"
        aiExplanation:
          type: ["string", "null"]
          description: "Human-readable explanation of why this category was chosen"
        aiTags:
          type: ["array", "null"]
          items:
            type: string
          description: "Semantic tags assigned by the AI classifier"
        aiSuggestedDestination:
          type: ["string", "null"]
          description: "Suggested folder/location for organisation (display-only; never applied automatically)"
        aiSuggestedAction:
          type: ["string", "null"]
          description: "Short human-readable description of the AI's suggested action"
        aiProvider:
          type: ["string", "null"]
          description: "Identifier of the AI provider that produced this classification"
...
    SearchFilters:
      type: object
      properties:
        path:
          type: ["string", "null"]
        extension:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        aiCategory:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        riskLevel:
          type: ["string", "null"]
          enum: [low, medium, high, critical, null]
        minSizeBytes:
          type: ["integer", "null"]
        maxSizeBytes:
          type: ["integer", "null"]
        dateFrom:
          type: ["string", "null"]
        dateTo:
          type: ["string", "null"]
        scanId:
          type: ["integer", "null"]
        duplicatesOnly:
          type: boolean
    SearchResults:
      type: object
      required: [query, filters, explanation, findings, total]
      properties:
        query:
          type: string
        filters:
          $ref: "#/components/schemas/SearchFilters"
        explanation:
          type: string
          description: Human-readable description of how the query was interpreted.
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        total:
          type: integer
    SearchHistoryEntry:
      type: object
      required: [id, query, filters, resultCount, createdAt]
      properties:
        id:
          type: integer
        query:
          type: string
        filters:
          $ref: "#/components/schemas/SearchFilters"
        resultCount:
          type: integer
        createdAt:
          type: string
    SavedSearch:
      type: object
      required: [id, name, query, filters, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        query:
          type: string
        filters:
          $ref: "#/components/schemas/SearchFilters"
        createdAt:
          type: string
        updatedAt:
          type: string
    SaveSearchInput:
      type: object
      required: [name, filters]
      properties:
        name:
          type: string
        query:
          type: string
        filters:
          $ref: "#/components/schemas/SearchFilters"
...